+++ EcosystemSimulation/src/io/tashtabash/sim/space/resource/dependency/TemperatureMin.kt
@@ -8,7 +8,7 @@ import kotlin.math.sqrt
 
 class TemperatureMin(threshold: Int, deprivationCoefficient: Double) : Temperature(threshold, deprivationCoefficient) {
     override fun satisfaction(tile: Tile, resource: Resource, isSafe: Boolean): Double {
-        var result = (threshold - tile.temperature).toDouble()
+        var result = threshold - tile.temperature
         result = 1 / sqrt(max(0.0, result) + 1)
         return result
     }
